Refactor subscriptions to balances
